Achievements now show up in the xlogfile
They appear at the end of the line, as tnntachieve0=0xaaaa... and
tnntachieve1=0xbbbbb... If we somehow manage to overflow 128 distinct
achievements, this will naturally flow over into tnntachieve2 and so on.

Achievements, in the order they are defined in you.h, correspond to the
least significant bits of these achieve flags. Examples:

Achievement #0 is tnntachieve0=0x1
Achievement #1 is tnntachieve0=0x2
Achievement #63 is tnntachieve0=0x800000000000000
Achievement #64 is tnntachieve1=0x1

... and so on.