New #foodseaten command, fix major bug in food achievement
The foodseaten extended command brings up a window that shows you which
food objects you have and have not yet consumed. This is intended for
tournament use, not just debug use.

The bugfix was that a player who ate *anything* would get the all-foods
achievement, and a player who ate any produce would get the all-produce
achivement.