Fix: bad check for shapeshifter in A_TAMED_SHAPECHANGER
Previously it was only working when the shapeshifter was in an
un-shifted form.

With this new check, anything that is currently using cham OR any
shapeshifters in an un-shifted form are valid for this achievement. This
will also allow things like shifted vampires to be fair game (but
they're shapechangers, so, it's fine.)