Implement extensible "TNNT globals" storage
In order to prevent tacking more and more things onto struct you. This
creates a new global variable "tnnt_globals", which is a struct that
holds TNNT specific globals. The reason for doing it this way is that
TNNT globals will be saved along with the game, allowing persistent
storage in-game, without needing to modify the save/restore code again
every time we might want to add more globals. It also helps put all the
TNNT modifications to globals in one place.