Add the scroll of missing code
This is a scroll that is a macguffin: it's worth nothing, isn't
inherently magical, has a nonrandom appearance, isn't writable, won't
vanish when read, and only produces cryptic messages when you try to
read it or write it. When the game begins, three copies of it are sent
to random levels in the Dungeons of Doom. Whatever could it be???