Replace #achievements with #tnntdebug, controlled by TNNT_BETA flag
TNNT_BETA is a config option intended to enable things friendly to
public beta testing. Removing its definition will either disable these
features entirely or make them accessible only to wizard, depending.

Currently the only such thing behind this ifdef is the new #tnntdebug
command, which if TNNT_BETA is not defined will be a wizard-mode-only
command. #tnntdebug is a more advanced version of #achievements; it pops
up a whole window to display a lot of information at once. The
achievements, still in hexadecimal, are shown in this window; the status
of the devteam quest and the levels on which the scrolls originally got
sent to are displayed there.