Implement the Devteam Quest Leader
This adds special behavior for a devteam member named "Mike Stephenson"
to discuss a special quest with the player when #chatted to. Upon
completion, he will reward the player with the Really Cool Shirt.

Adds two com_pager messages for the long form description and
congratulation.

Note that this commit doesn't actually add a Mike Stephenson monster to
the devteam level, there are still some issues which would be best
addressed in a separate commit.