Hack fix: devteam stairs didn't generate if quest portal was on level
This is a bad, bad hack, and would properly only be resolved by
supporting two or more branches on a single dungeon level in vanilla. I
don't have the ability or knowledge to add such support right now, so we
have this hack instead. It consists of two parts:

1. When a branch is placed, before finishing, it sees if there are any
other matching branches on that level and recursively calls place_branch
until there are no more.

2. Comment out an impossible designed to fire if the game detects two
branches on the same level (this only affects overviews though, as far
as I know).