Begin implementing #tnntstats command
Unlike #tnntdebug, this is to show an ordinary player statistics
regarding the TNNT achievements and other custom content like the
devteam quest. Currently it just shows the number of scrolls returned to
the devteam (only if the quest is started) and a notice that you need to
do #snacks to get the foods eaten summary.