Move all TNNT achievements definitions and globals into decl.h
Globals now are in the tnnt_globals struct; at the time, it made sense
to stick achievement related stuff in u_achieve, but now that we have
tnnt_globals that's a better place for them. Because tnnt_achievements
is no longer in you.h, the list of achievements has been moved into
decl.h as well, along with the tnnt_achieve macro.

Fix all references to u.uachieve.foo to point at tnnt_globals.foo.