Fix bug with taming 7 felines
Previously it just ticked up every time a feline was tamed. This allowed
you to let a f pet untame and then come back and retame it for the
achievement. Now, store an array of 7 monster ids in the tnnt globals
struct, and don't count it as taming a monster again if the m_id is old.
This should be secure against reanimating a pet corpse; the id should be
preserved in that process.