Add achievement names and descriptions to game
Finally going forward with this. This does two main things: it lays the
framework for being able to check your achievements in-game, and
provides the basis for the game source being the single point of truth
for the achievement names and descriptions on the web backend, and
allowing us to avoid nasty synchronization problems.