Implement in-game achievements display
This is the single biggest requested feature from the 2018 tournament: a
way to check which achievements have been earned or not in-game. This
re-adds the #achievements command as a non-wizmode command which simply
dumps all the achievement names, descriptions, and earned/unearned
status out at once. The user can choose to see only earned ones, only
unearned ones, or both.

Currently in the tty windowport there is no way to page back through the
list, similar to how the discovery list works; not sure how to fix this.

Achievements counted at the end of the game will currently show as
unearned in the entire game whether or not they are actually earned; the
plan is to move the triggers to when the player actually earns them so
that they will show this correctly.