Trigger achievements NOT at end of game whenever possible
There were several only being recorded in end.c at the end of the game,
which was all well and good when the only achievement record that
mattered was the xlogfile. Now, however, since the player can view their
achievements in game, it would be misleading to not call tnnt_achieve on
them at the moment they were earned.

The behavior of a bunch of "kill things" achievements was changed
slightly: because the triggers live in xkilled now, the player needs to
kill the thing themselves, rather than just having the thing die to
possibly totally unrelated circumstances. This is weird for nazgul and
erinyes: the player only needs to kill the last one; the others can die
however. More discussion is probably needed.

Two achievements, finishing a game with 50K or 100K gold, remain in
end.c, since by definition they can only be triggered at the end of the
game. Thus, they will continue to never show as achieved in-game.