Partial conduct achievements now require you to get assigned the quest
As was demonstrated last year, requiring you only to reach your quest
leader was more a test of how lucky you could be starting an
Archeologist and digging down until reaching the quest and getting to
the leader without dying, rather than actual skill at playing a conduct.

So, the requirement has now been changed to get to the point where
you're actually given the quest, meaning that on top of reaching the
quest leader you must be experience level 14 and piously aligned. You
actually need to play the game with the conduct intact for long enough
to get to that point.