NPC Deathmatch branch (initial). Draft submission for the NPC deathmatch branch and corresponding map.
Idea behind map design - the player enters the branch via a portal, and they find themselves in a small foyer with the north and south walls lined with statues of players who have ascended before them. As the player makes their way west down the hallway, there are two messages burned into the floor giving a hint as to what to expect, and another right before a closed door stating this is the point of no return. Should the player go through the door, they're committed to what comes next.

The main arena past the door is where the player meets an NPC that is a close copy of the last real player to ascend. Here is where they fight to the death. I placed some columns in the room to provide a type of cover from ranged attacks. A vertical split moat in the middle of the arena provides a bit of a bottleneck when trying to go across the arena. Added a few random spiked pits that can affect both the player and NPC.

This map is just a draft submission - this will almost assuredly be edited a fair amount before it's finished. The hints/linux file was edited to make it simple to compile and setup a wizmode build for testing.

Some things that defintely need to be changed:
* I put the portal on the same level as the sokoban stairs strictly for ease of access during testing. This will not be its final access level
* The random traps should be restricted to the main arena only and not appear in the foyer
* I'm not sure how to make the statues in the foyer be those of past players