Implement killed-groups-of-monsters achievements
This implements the 5 "kill a member of X number of species"
achievements and the Learn the Alphabet achievement.

The game now tracks monsters that you, personally, killed, in addition
to those that just died some way or another. Only monsters that you kill
personally count towards these achievements; pets or conflict don't
count. Side effect: the Nazgul and erinys achievements now require the
player to kill all of them instead of just the final one.

New function tnnt_common_monst determines whether a given monster is
eligible for these achievements (generally, this pool consists of
monsters that have a frequency greater than 0, with some adjustments).

New command #species shows the species that have and have not had a
monster of that type killed yet, in addition to showing the letters
killed for Learn the Alphabet.