Seven new conducts
Promotes survivor to a conduct recorded in the xlogfile, and adds
others. Also adds trackers to the #conduct command so you can check
their intactness.

All seven are recorded as bits in the xlogfile's "conduct" field:
0x02000: survivor
0x04000: never took from a swap chest
0x08000: never let the quest nemesis die
0x10000: never let Vlad die
0x20000: never let the Wizard die
0x40000: never let any high priest die
0x80000: never let any Riders die

Adds another field to uconduct (for tracking swapchestless); as such,
this is a savebreaking commit.