Two additional conducts: artifactless and petless
This adds petless and artifactless (specifically, "never touch an
artifact") conducts. They are recorded in the conduct field of the
xlogfile:
artifactless: 0x100000
petless:      0x200000

Also adds trackers to #conduct for them. This adds the two new fields to
uconduct, so this commit breaks saves.

The specifics of petless are taken from xNetHack's implementation. In
the interest of not forcing petless players to torpedo their own
alignment record or generate conflict before reaching the Astral Plane
and receiving a tame guardian angel, I have ported the part of the code
that checks for petless and won't tame the angel if petless remains
intact when arriving on Astral. The only players that could be affected
by this would be ones that start the game with no pet and play the
entire game with no pet, but are expecting the guardian angel to appear
on Astral. I doubt any such players exist.

However, the xNetHack feature that neuters the taming effect of magic
traps into pacifying is not brought over; a petless player can still
accidentally break conduct by triggering a magic trap and getting
unlucky.