Two more conducts: permahallu, permadeaf
These are roleplay options, or "birth options" to use the NetHack4
terminology, and like blind and nudist conduct, require the user to put
OPTIONS=hallucinating or OPTIONS=deaf in their configuration file to
play a game with them active.

Permadeafness cannot be broken by any means. Permahallucination can be
broken, permanently, with an artifact that confers hallucination
resistance (Grayswandir), just like with the Eyes of the Overworld
breaking permablind conduct. There is a YAFM added to replace being
"shocked back to your senses" by a potion of sickness; this does not
affect permanent hallucination.

They are recorded in the conduct field of the xlogfile:
permadeaf:  0x400000
permahallu: 0x800000

Add trackers to #conduct for these, as usual.