Sundry achievement tweaks
"Dungeon Economics" - this could be triggered by selling something also,
so update its description accordingly.

"Money Talks" - you now need to bribe with 25000 or more gold to earn
this, rather than 1.

"One Hit KO" - up the damage required to 20; put this in a TNNT_OHKO_DMG
constant so that if the damage threshold ever gets changed later that's
the only place it needs to be changed. Also remove some cases that could
never trigger this achievement because they could reach 10 damage, but
never 20.