Refactor #tnntdebug
It now gives a menu offering various debug commands you can do, in order
to prevent having a multiplicity of debug-only commands. The 's' option
gives the same hidden-from-normal-play information that it did before,
and the new 'w' option writes your current player state to the npcdata
file as if you were ascending.