Implement function for spawning in an NPC monster
This is the function which reads from the 'npcdata' file and uses that
information to create the NPC player monster in game, with all possible
fidelity to the original player.

Contains some adjustments to prevent the challenge getting too hard or
too cheap: among other things, the HP maximum can never be below 200 or
above 500, and the level is always at least 14. If there's some error
reading the data file, it impossibles and falls back on creating a
generic Astral-level player monster.

I added a new command to the #tnntdebug menu which executes this
function and spawns the NPC in next to the player. Currently the
function isn't called from anywhere else (the arena is still empty).