Do not pass object owornmasks to the NPC at all
This was causing some bugs (that I noticed: npc could spawn wielding
something the ascender had been wielding, and then wield something else,
now apparently wielding two weapons. There were probably other bugs
too.)

I also realized that if the AI considers it suboptimal to be wearing the
gear the ascender was wearing, it will spend its turns re-equipping
itself - not what we want for the opening salvo of a deathmatch. This is
the main impetus behind removing owornmask, since if we're going to just
call m_dowear and let the AI pick the gear it wants anyway, it doesn't
matter what the ascender had.