Track state of deathmatch
Adds two flags to tnnt_globals. One flips on when you enter the arena
(prompting a challenge message and waking up the NPC the first time you
do) and one flips on when the NPC is killed. #tnntstats now shows the
status of the deathmatch if you have already started it.

To track the NPC uniquely, we use the 'mcloned' monster field, which is
never used for player monsters. As a bonus, setting mcloned prevents
death-drop and corpse drop.