Exploit closures: branchporting
Takes care of branchporting out of the arena while the match is in
progress (this couldn't be allowed even if it were acceptable as an
escape because it wouldn't have the necessary information to destroy the
magic portals).

Also blocks the Arena from being put on the branchport menu if you have
fled the match, so no branchporting back in to resume the match.