NPC inventory for deathmatch - initial commit. When the NPC for the deathmatch arena is created, they are assigned ascension-level gear. The role they assume will dictate what kind of gear is received (armor and weapons [melee and ranged]), as well as an amulet and some lizard corpses.
Some logic has also been added so the NPC will decide which weapon to
use first (melee or ranged) - as the NPC is a good distance away from
the player once they enter the arena, the NPC will use its ranged attack
when closing with the player.

Current issues to resolve:
* NPC will spawn with both whatever armor/weapons/amulets it was
wielding or wearing when its template was created, along with these gear
assignments. It's supposed to be either/or (seems the NPC
is created before the !which_armor() lines kick in)
* Lizard corpses are made, but appear at the NPC's feet and not in
inventory. Not quite clear on how to get them into the NPC's inventory