Exploit closure: only the player can use the deathmatch portal
This prevents you from bringing in your personal army of pets to fight
for you (and also closes off some loopholes for toting the NPC's gear
out of the arena before the match ends). Note that you can still summon
pets inside the arena, but you won't be able to take them back out
again.

This also blocks followers following you through the portal. I haven't
tested what happens with the Wizard of Yendor, but what ought to happen
is that he will not follow you through the portal but can later appear
inside with "So thou thought thou couldst elude me, fool".

Monster levelporting out of the arena appears to be blocked already by
virtue of it being a single-level branch.