Deathmatch reward: now one item of your choice from the NPC
Adds a flag to struct obj that tracks items as 'transient'. Every item
the NPC gets is flagged with this. Transient items cannot be stolen by
you; they turn to mist in your hands instead. They will also not merge
with nontransient objects.

Upon death of the NPC, all transient items on the level get removed from
wherever they're hiding and placed on the spot the NPC dies. Upon
subsequent addition of one of these items to your inventory, they get
removed from wherever they are, again, and deleted (minus the one you
just got). Warning: nontransient objects inside a transient container
will be deleted, too.

As an artifact of earlier code iterations, this also tracks the object
ID of the thing you got from the NPC as a tnnt_global. This isn't used
for anything currently but I'm leaving it in in case we decide to do
something with it.