Exploit closure: awakening the NPC from outside the arena
Various things that set msleeping to 0 now awaken the NPC and begin the
deathmatch even if you haven't stepped across the doorway. I didn't do
this for all effects that set msleeping to 0, only the ones that seemed
likely to be triggered with the NPC.

Initially I was going to implement it such that effects that awaken
monsters at a distance would have no effect on the NPC, but decided that
it jumpstarts the deathmatch instead. The important state that needed to
be prevented was having the NPC be awake and coming for you before the
game recognized the deathmatch as in-progress.