Implement #achievements view for vanilla achievements
These are the standard 11 we count as achievements in TNNT: get the
Bell, get the Book, enter Gehennom, etc. This only affects the output of
the #achievements command - a new tnnt_achvmt_data structure is
introduced to track the vanilla achievements, but no existing data
structures are changed and nothing new happens with the xlogfile.

We may decide to merge these into tnnt_achievements in the next
tournament. Currently they're handled in a separate loop and output as
e.g. "#V05".