Fix: guardian angel was still made tame unconditionally
I apparently forgot to delete that line after copying it into the if
(u.uconduct.pets) block. This gave you a tame guardian angel, but did
not break petless - best of both worlds. Delete the duplicate line
outside the if so the intended behavior of only taming the angel if
petless conduct has been broken.