Remove moats from the NPC deathmatch arena.
Reason - it became ridiculously easy to kill the NPC on this level by
simply freezing part of the moat, get the NPC to walk on the ice, then
melt the ice underneath the NPC, causing them to drown. The easiest
solution to this without changing any game mechanics is to
remove/replace the moats (in this case, they are replaced with lava
pools). Ideally I'd like the moats to remain and add some logic so that
the NPC avoids walking on ice, or they can already fly/levitate, etc.
But that can be a discussion with the team for another day.