Minesweeper requires disarming the final non-player-made mine
Saw a bit of an exploit here: what the code was checking was that the
player was disarming the last existing mine in Fort Ludios. That meant
that you could potentially summon a swarm of monsters into the treasure
room, let them rampage a bit and blow up all the mines (which doesn't
set the player-blew-up-ludios flag that makes you ineligible for the
achievement), then set one of your own land mines and disarm it to earn
the achievement.

What changes here:
  - The check is only made when you disarm a non-player-made land mine.
  - Only look for any more non-player-made land mines on the level. If
    the player sets one of their own and it doesn't go off by the time
    they get around to disarming the ones in the vault, that shouldn't
    block them from earning the achievement.