Fix: doppelganger -> gremlin -> mplayer counted as deathmatch opponent
The assumption that the deathmatch opponent is unique in being a player
monster with mcloned = 1 proved false. A doppelganger can become a
gremlin and multiply in water, producing more doppelganger-gremlins that
now have mcloned = 1. These can turn back into player monsters, at which
point they count as the deathmatch opponent. The effect was that things
which cause a monster to wake up were starting the deathmatch on random
levels where such a doppelganger happened to be (and making the real
deathmatch effectively unwinnable).

Fix this by making the system more robust - now the deathmatch
opponent's monster id is tracked like the quest leader's is. This
augments tnnt_globals and is thus a savebreaking change.