Fix: foods_eaten was only checked after eating corpses, tins, or globs
I'm not sure how this happened. This effectively meant that you could
eat all the food types to earn Smorgasbord (and #snacks would show it as
such), but you wouldn't actually earn the achievement until you next ate
a corpse and the foods_eaten bitfield was checked. The fix is to add it
to the one other place in the code foods_eaten is updated so that the
achievement is checked after eating any food item.