clarify the arena deathmatch loot
several players during TNNT and in the post-tourney survey were surprised
when all the items vanished.  instead of all the items mysteriously vanishing,
re-use the booming announcer voice from tnnt_globals.deathmatch_started

Note by copperwater: adjusted this commit a bit. It now uses cansee()
rather than !Blind (maybe the opponent got taken down with a stinking
cloud, or died behind a wall to a pet, and can't be seen), and I kept
the original "coalesce" message for when the pile is visible. Also, a
bit of TDTTOE: if there's only one transient item, none of the pile
messages appear, because there is no pile.