Fix: crash during restore (hallucination resistance).
This was very recently fixed in NetHack 3.7 git commit 9f230c3, but was
initially discovered from a crashed game on NAO (3.6.6) that could not
be restored. The fix here is much more simple than the fix for 3.7, but
here's the gist of it:

'Triggered by Grayswandir's hallucination resistance.  If the game
is saved while hero is hallucinating but having that be suppressed
by wielding Grayswandir, is riding, and the steed is on an object,
then during restore the hero's location will be updated because
of the presence of the object but the attempt to display the hero
there is made before u.usteed has been restored and fails.'