Fix an unexpected issue with Bar Brawler achievement
This is apparently a latent vanilla bug, or if not actually a bug then
unstable code. When hitting a monster with a potion, the potion gets
destroyed, but hmon_hitmon still thinks obj is valid. Bar Brawler and a
bunch of other "kill a monster in this circumstance" achievements were
therefore all reading freed memory when the kill was made with a potion.

I did not attempt to fix the underlying issue because I think setting
obj to NULL in this case might cause other things in hmon_hitmon to
segfault. Instead this is a somewhat clumsy workaround.