Remove "Implemented/tested" comments from achievement enum
All 2021 new achievements are implemented and tested; I think it's a
safe assumption that the ones from previous years are too, even though
some of them never got the "tested" added to them. This was just rather
a hassle to maintain and forced a full rebuild every time it was updated
because it's in decl.h.