Reorder the achievements
Over time the practice of adding achievements as they get dreamt up has
created some oddities, like "why isn't Mineralogist with all the
'identify all $objectclass' achievements". This aims to fix that by
collecting thematically similar achievements into chunks. Not yet
determined if the web frontend will display these chunks; there isn't
any code here in the game for treating them as part of a chunk, so
probably not.

There are also a number of mini-chunks floating around that didn't get
specifically mentioned with a comment. Generally, if a group of
achievements only had 4 or fewer, I didn't bother with it, but I did put
some effort in to make them consecutive.