Block stack splitting for items in swap chest
Taking a partial stack from the swap chest causes all sorts of segfaults
due to the way swap chest items currently work (obj->swapobj_filename
and obj->ocontainer are mutually exclusive, OBJ_CONTAINED vs OBJ_INSWAP
needs to be accounted for in splitobj)... and because removing an item
from the chest deletes the associated file no matter what, it's not as
though removing a partial stack would leave the remainder for everyone
else.

I do think it would be good to revisit this in the future, to accomodate
splitting of swap chest stacks properly, but it will probably be a
reasonably involved project.