Fix: crash entering Astral
If the player had no pets when entering Earth, the game could crash when
entering Astral, apparently related to code of the form

    planes_pet_m_ids = alloc(num_pets); /* num_pets == 0 */
    /* ... */
    free((genericptr_t) planes_pet_m_ids);

The C standard suggests the result from malloc(0) should be safe to give
to free(), but for some reason this is causing problems -- maybe the
difference between malloc() and NetHack's alloc().  I wasn't able to
reproduce it myself so it may depend on the OS/standard library/etc, or
maybe some compile-time settings for NetHack that change how alloc()
works... anyway, preventing the alloc() and free() unless num_pets is
nonzero seems to fix it, whatever is causing it.