Future fix and a current shim for planes pet saving
What I failed to account for in implementing the Tenacious Companion
achievement is that dynamically allocated memory can't just be saved
along with the rest of the big tnnt_globals blob. Unfortunately, since
the tournament is currently running, we can't fix this by saving that
data properly, since it will obviously break saves.

So this commit provides the functionality we're going to want to have
for 2022, locked behind an ifdef, and in the meantime it provides a shim
to keep the achievement as attainable as possible in the current
tournament. This is done by zeroing out num_planes_pets and
planes_pet_m_ids on a *manual* save, which will prevent bad pointers
from getting read into the save file. If the game is never saved between
entering Earth and Astral, those are not zeroed so the achievement is
still attainable.