Vanish devteam when message says so
This is a pretty insignificant little touch, but when you kill a member
of the devteam, there's a line describing the remaining devs 'vanishing'
from the dungeon -- yet they hang around after that anyway (for all of
the two or three lines of text a normal player would see before the game
ends...).  Actually remove them from the game once that message has been
displayed.