Fix #25: Beast Master granted for hatching egg
Beast Master ('tame a monster without using food') was given for things
like hatching an egg, using #monster in a wereform, cutting a tame long
worm in half, and other scenarios which aren't really taming an existing
monster (at least not from the player's perspective).  Move granting the
achievement out of tamedog(dog.c), which can't distinguish between these
different uses, and instead grant it specifically for using a magic
harp, taming scroll, or charm monster spell, or for the taming effect
from a magic trap or the Book of the Dead.

Fixes #25