Implement "Hoist With Its Own Petard" achievement
For having a monster kill itself with a wand zap.  This is a pretty
simple implementation.  Is there anything else that could trigger these
conditions (monster is dead immediately after zapping its wand)?  I
guess if the zap hits a gas spore and the explosion kills the monster:
is that a problem?  I don't know...