Change 9-Nazgul-kills achievement to "collect all 9 Nazgul rings"
The spirit of this achievement, to kill all 9 Nazgul, was easily
bypassable by repeated stoning and unstoning of a single Nazgul. So,
change it to require having each of the 9 rings enter your inventory at
some point. (You don't need to have all 9 in inventory at once, but
anyone who does will have earned the achivement.)

Also adjust a few bits of code with the other "collect an item"
achievement.