Add "Po-Loew-morph" achievement
For creating a golem by polymorphing a pile of objects.  Another name I
came up with and am using pending a better suggestion.  It's a reference
to Rabbi Loew, the Maharal of Prague, famous for creating the Golem of
Prague (among other things, like his great scholarship and study of the
Torah).  It's an especially bad stretch, even by the standard of the
other achievement names I've come up with.