Refine Right Back At Ya to only include indirect boomerang hits
The impetus for this one is because people were earning the achievement
just by chucking a boomerang into a monster directly in front of them.

Initially I implemented this just by checking whether the monster was
lined_up with the player, but then realized there are plenty of places
along a boomerang flight where a monster is lined up with the hero that
aren't straight in front of them.