Fix: Prevent entering the Deathmatch Arena with unique items
Because if you do enter, you could drop those items then flee the
deathmatch so that it gets sealed off, and then the game is unwinnable.

Initial patch is by qt but it only covered the regular case, but this
needed to be closed off with branchport as well as regular portal
traversal, so I extended it to that, and also made it so you can take
any items in if the deathmatch is done, or take invocation items in if
the invocation is done.