Fix: better way of handling livelogging of destroying a player's ghost.
TNNT already livelogged the destruction of another players bones ghost
(see commit 18fee71), but that method added a whopping 25 bytes to
struct monst, which is unnecessary. This commit lowers that amount of
extra bytes to something like 6-7, and also allows for livelogging not
just bones ghosts, but any form the player might arise as (vampire,
wraith, etc).

Thanks paxed for the initial code.