Fix: don't stringize ach. names/descs for YAML
This caused achievement descriptions that were broken up into more than
one constant (like "line 1 "\n"line 2") to be preserved this way,
pre-concatenation, in the resulting YAML file. To avoid this, leave
everything unquoted, since that's allowed in YAML (trying to add quotes
some other way would be complicated, since it would require escaping any
double-quotes used in an achievement description).

This means that names/descriptions starting with a '#' character are
interpreted as comments, so for now at least just rephrase the one
achievement description that fell into that category.