Refactor and add progress indicators for ID-a-set-of-items achievements
The code may look clean now, but I assure you it went through several
ugly iterations before reaching this point. In the end, though, I was
able to figure out how to house all information about achievements for
identifying object sets in one place, and use that for both checking
whether an achievement is earned when something new gets discovered, and
showing a little "x/y" summary in the discoveries list, even for the
trickier classes like gems which have multiple achievements.

Originally I was going to put the summaries in #tnntstats but the
discoveries list is both a more natural place for it, and allows all the
code to stay in one file.

One small adjustment to one of the achievements: it is no longer
required to formally identify water for the potions achievement, since
blank scrolls and spellbooks weren't necessary either.