Add achievements to end-of-game disclosure
The prompt can be hidden, shown, etc, by including 't' in the 'disclose'
compound option, similar to other end-of-game disclosure questions
(e.g. 'OPTIONS=disclose:+t' will always show achievements, 'yt' will
default the prompt to 'yes', etc).  't' for TNNT, since 'a' was already
taken.

Also, set the size of tnnt_globals.tnnt_achievements 'automatically'
based on the number of achievements.  I think this is OK since it is an
"integer constant expression" but can always be reverted if it causes
problems.