Add infrastructure for reporting achievements from an in-progress game
This is a pretty popular request - players would like the scoreboard to
show them achievements that they have as soon as a game is finished, but
there was previously no mechanism by which the game could report on
achievements while not yet having an xlogfile entry.

This creates that mechanism, a simple writing of files to a new
"temp_achievements" directory which record unambiguously which game it
refers to and which player, and the tnntachieve fields. These can then
be collected and placed in a location the web server can read from.

Biggest change here: tnnt_achieve changes from macro to function, since
gaining a new achievement is what triggers writing the file.