Fix: "Recharged" granted when charging fails
The "Recharged" achievement was granted if you used a scroll of charging
on any object, even if the object was not chargeable or the scroll was
cursed and caused the object to be discharged (or had no effect on spe
at all).  Try to avoid granting the achievement the scroll doesn't
actually have a charging effect.  Overcharging a wand so that it
explodes is considered to count.

Co-authored-by: Joanna Irina Zaitseva-Kinneberg <jjadoyle@gmail.com>