Randomize swapchest item destination server
When an item is put into a swapchest, it will go randomly to one of the
three servers (us, eu, au), not just the one it originated from.  This
is intended to balance out the swapchest inserts to benefit servers with
fewer players, as an alternative to making all three servers share a
swapchest (which was too laggy, and alternative implementations
involving periodic resynchronization ran a risk of being exploitable by
allowing duplication of swapchest files).

This is experimental and might be rolled back if it doesn't work well.
Relies on a separate cronjob running on each server that will
periodically move swapchest files to the destinations they are prefixed
with.