Track item donors when someone takes an item from a swap chest
This is intended to expose this data to the server so that we can do
some interesting stat tracking like "who is the best/most generous donor
of items people want to use" - the idea is that people who put in useful
items that other people end up taking out will show up on this list more
often than people who put in junky, useless, but still acceptable items
like scrolls of amnesia.

This is behind defines for TNNT_SWAPCHEST_DIR, so if TNNT is built with
the other swapchest stuff disabled, the file won't ever exist.