Make most of the devteam office items less than 100% chances
Over the past couple tournaments, we've noticed some unwanted patterns
developing with this level and even fielded some feedback in the surveys
along the same lines. The Office contained a lot of guaranteed items and
monsters, and this had a noticeable effect on achievements (about a
dozen achievements could be trivially obtained by anyone who gets to the
level, including rarer ones like the one that requires a grappling
hook,from the stuff lying around) and even overall game balance (a
guaranteed spellbook of detect monsters, for instance, can be
game-changing for many characters).

So this commit sticks mostly low percentage chances on most of this free
stuff. Enough there for the level to remain not empty feeling, but
solving most of the problems with players treating them as a dependable
store of resources.