Do not reward finishing the devteam quest when player killed Izchak
In such a case, the player will still be awarded the achievement for
finishing the quest, but will not get the Really Cool Shirt. The player
must have been personally credited for Izchak's death.

I doubt anyone is actually going to hit this case intentionally, but
it's a feature idea that's kicked around for a while.