Make the swap chest only show other people's items
This addresses a problem that was raised by several people: you could
insert an item which you had previously called the type of, or which you
had duplicates of, and then look inside the chest to see it fully
identified, thus getting a free identification of that type. The items
being assigned to random swap chests partly addressed this in that it
was only 33% reliable, but 33% is still a lot more than 0% (and also if
TNNT is ever run on a single server, it would have done nothing to
help).

Initially the plan was not to show any items that the player had put in
since the start of their own game, but since that would have required
changing the swap object file format I decided to simplify the
implementation and only show items from other people.

In addition, I took the opportunity to do some more refactoring of the
swap chest code: the name written into the file is now just the player
name, with the "a token from" and other prefixes being added when
someone else opens the chest. There's no longer a correspondence between
the prefix and whether it was someone's 1st, 2nd, or 3rd donation. Also
addressed a lot of formatting issues with an if statement or for loop
with a single line of code all on one line, which are hard to read.